requirements.txt
# Market data
yfinance>=0.2.36
fredapi>=0.5.1

# LLM integration
openai>=1.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
freezegun>=1.4.0
pytest-mock>=3.12.0

# Utilities
python-dotenv>=1.0.0

# AI Agent Framework (Phase 1)
pydantic-ai>=1.4.0
mcp>=1.19.0
